requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
flask==3.1.0
flask-sqlalchemy==3.1.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
jinja2==3.1.6
jsonpath==0.82
loguru==0.7.2
markupsafe==3.0.2
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
sqlalchemy==2.0.40
typing-extensions==4.13.1
urllib3==2.3.0
werkzeug==3.1.3
